requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyjwt==2.8.0
python-dotenv==1.0.0
pytest==9.0.2
httpx==0.23.3
